Open --exec subprocesses with /dev/null as stdin
This is necessary, because when we --exec SSH, it reconfigures the
terminal at stdin which results in signal handling not working
correctly. Probably, the terminal is switched to raw mode so that SSH
can send things like Ctrl-C to the remote side. This is fine except
for the situation, when multiple processes use the terminal as
stdin (which is our case). Then, Ctrl-C can be eaten by other
processes than SSH and the key is netiher sent to the SSH remote side
nor does it generate SIGINT locally due to reconfigured terminal.

One more fix to #17.