Run --exec processes in background process groups
This fixes the problem when an --execed process keeps running
indefinitely after the benchmark is terminated.

We partly revert 1003e9d ("Do not block signals in child processes",
2020-03-10). There is no other easy way to (almost) reliably kill both
subshell and its child processes. In the mentioned commit, we didn't
want ssh subprocesses to be blocked by reading from stdin terminal,
which is not allowed for background processes. This is no longer a
problem, because the subprocesses is run with /dev/null as stdin.

We also add a test. Before these changes, the test failed.