Run the benchmark in a background process group
This allows killing all processes (the process group) that are a part
of the benchmark, hopefully solving the problem of thermobench waiting
indefinitely for benchmark termination in some cases. This may be a
better solution than #32.

The disadvantage of this approach is, that the benchmark cannot read
from terminal via stdin, because background processes cannot do this.
Currently, we don't have any benchmark that needs this, so it should
be no problem.