Get rid of memory leak in benchmark stdout processing
It was not memory leak per se, because all memory was released at the
end, but it was sufficient to consume all memory during long
measurements.

The problem was in always allocating new FILE via fdopen() in
child_stdout_cb. Unfortunately, it could not be fixed by simply
calling fdopen() once at the beginning, because then we were not able
to reliably determine feof() and thermobench failed to terminate quite
often. The reason was that FILE* is not made to handle non-blocking
file descriptors well.

To fix the problem, we removed the use of FILE* and implemented proper
buffer management ourselves.

Fixes #37