benchmark/CPU/instr: Allow running demos-enabled version without demos
Requiring running the benchmark under DEmOS would require the user to
recompile the programs when they want to run some benchmarks without
DEmOS. This would be annoying. Instead, allow the benchmarks to be
used in the same way as before (without DEmOS), but emit a warning
that DEmOS was not found.

Also get rid of #ifdef to prevent "‘demos_enabled’ defined but not
used" warning.