sensors.imx8: Change how thermocamera data is accessed and add heat sources column
Previously, we polled directly the camera. The problem with this
approach was that measurements on both our boards contained the
thermocamera data, but the camera measured only one board.

Now, we use SSH to access the data and the SSH account is configured
to return the data only when connected from the "correct" board. For
the other board, no data is returned (the SSH runs /bin/cat).

Also, the returned data contain not only the information about
on-board temperatures, but also information about detected heat
sources.