benchmark/cl-mandelbrot: Make more suited for thermal  experiments
One of the goals is to have a long running kernel which runs which
runs for the same amount of time in all configuration. For this
reason, the default value of escape_radius is ∞ and the default value
of max_iter is increased to 65536.

We also increase the default image with to match the default global
work size and decreased image height. We don't care about the
resulting image, but do care about low memory bandwidth (this should
be compute-bound benchmark).