Add sensord tool
It reads all sensor values and publishes them for Thermobench as well
as Prometheus.

Currently, most of the configuration is hardcoded in the source code,
but this can be changed later, if needed.

It is planned that thermobench will connect via SSH by running

    nc -U imx8

i.e., it will connect to a UNIX socket, where sensord pushes sensor
values.